#175 moved order creator tests into test_evaluator_order_creator.py
